Re-add compatibility with RF 3.1.2 and earlier
